refactor: extract backup helpers from resumes.ts into lib/resumes_backup.ts
Move 3 types and 11 backup-related functions (projection, filtering,
normalization, sorting) into lib/resumes_backup.ts, reducing resumes.ts
by ~128 net lines. All symbols re-exported for backward compatibility.
